+++ expense-tracker-server/src/main/resources/db/migration/sqlite/V1.0.0__create.sql
@@ -1,4 +1,4 @@
 create table backups (id GUID not null, categories integer, expenses integer, persons integer, time datetime, primary key (id));
 create table persons (id GUID not null, name varchar(255), primary key (id));
 create table categories (id GUID not null, color_blue tinyint, color_green tinyint, color_red tinyint, description varchar(255), name varchar(255), primary key (id));
-create table expense (id GUID not null, amount float(19), date datetime, description varchar(255), category_id GUID, person_id GUID, primary key (id));
+create table expense (id GUID not null, amount numeric(19,2), date datetime, description varchar(255), category_id GUID, person_id GUID, primary key (id));
